remove files after successful wget

+++ gpgpassman/gpgpassman.sh
@@ -5,8 +5,8 @@
 # Also with 'zenity', you can execuite 'gpgpassman.sh dec' for direct access to decrypting passwords; can be used with a keybind.
 # Written by simonizor 3/22/2017 - http://www.simonizor.gq/scripts
 
-GPMVER="1.2.0"
-X="v1.2.0 - URLs for updating have changed due to adding folders to repo; if you have version 1.1.9, you will have to update manually :("
+GPMVER="1.2.1"
+X="v1.2.1 - Made update process smarter by checking to see if download failed before removing files."
 # ^^Remember to update this and gpmversion.txt every release!
 SCRIPTNAME="$0"
 GPMDIR="$(< ~/.config/gpgpassman/gpgpassman.conf)"
@@ -20,13 +20,35 @@ updatescript () {
 cat >/tmp/updatescript.sh <<EOL
 runupdate () {
     if [ "$SCRIPTNAME" = "/usr/bin/gpgpassman" ]; then
-        sudo rm -f $SCRIPTNAME
-        sudo wget -O $SCRIPTNAME "https://raw.githubusercontent.com/simoniz0r/UsefulScripts/master/gpgpassman/gpgpassman.sh"
-        sudo chmod +x $SCRIPTNAME
+        wget -O /tmp/gpgpassman$GPMVER.sh "https://raw.githubusercontent.com/simoniz0r/UsefulScripts/master/gpgpassman/gpgpassman.sh"
+        if [ -f "/tmp/gpgpassman$GPMVER.sh" ]; then
+            sudo mv /tmp/gpgpassman$GPMVER.sh $SCRIPTNAME
+            sudo chmod +x $SCRIPTNAME
+            sudo rm -f $SCRIPTNAME
+        else
+            read -p "Update Failed! Try again? Y/N " -n 1 -r
+            if [[ $REPLY =~ ^[Yy]$ ]]; then
+                runupdate
+            else
+                echo "gpgpassman.sh was not updated!"
+                exit 0
+            fi
+        fi
     else
-        rm -f $SCRIPTNAME
-        wget -O $SCRIPTNAME "https://raw.githubusercontent.com/simoniz0r/UsefulScripts/master/gpgpassman/gpgpassman.sh"
-        chmod +x $SCRIPTNAME
+        wget -O /tmp/gpgpassman$GPMVER.sh "https://raw.githubusercontent.com/simoniz0r/UsefulScripts/master/gpgpassman/gpgpassman.sh"
+        if [ -f "/tmp/gpgpassman$GPMVER.sh" ]; then
+            /tmp/gpgpassman$GPMVER.sh $SCRIPTNAME
+            rm -f $SCRIPTNAME
+            chmod +x $SCRIPTNAME
+        else
+            read -p "Update Failed! Try again? Y/N " -n 1 -r
+            if [[ $REPLY =~ ^[Yy]$ ]]; then
+                runupdate
+            else
+                echo "gpgpassman.sh was not updated!"
+                exit 0
+            fi
+        fi
     fi
     if [ -f $SCRIPTNAME ]; then
         echo "Update finished!"
